Add VCF Content Library Samples
JIRA: https://vmw-jira.broadcom.net/browse/VCFCON-15284

This PR adds the content library samples to create, update, and print
a content library that is backed by 1 or more regions. This PR does not
need to add any additional fields in the application.yml file. When running
the ContentLibraryExample it either re-uses existing Regions in the VCFA instance
or it creates new regions if non exist.